+++ src/types/observableobject.ts
@@ -7,6 +7,14 @@ import {IListenable, registerListener, hasListeners, notifyListeners} from "./li
 import {isSpyEnabled, spyReportStart, spyReportEnd} from "../core/spy";
 import {IEnhancer, isModifierDescriptor, IModifierDescriptor, deepEnhancer} from "../types/modifiers";
 
+const COMPUTED_FUNC_DEPRECATED = (
+`
+In MobX 2.* passing a function without arguments to (extend)observable will automatically be inferred to be a computed value.
+This behavior is ambiguous and will change in MobX 3 to create just an observable reference to the value passed in.
+To disambiguate, please pass the function wrapped with a modifier: use 'computed(fn)' (for current behavior; automatic conversion), or 'asReference(fn)' (future behavior, just store reference) or 'action(fn)'.
+Note that the idiomatic way to write computed properties is 'observable({ get propertyName() { ... }})'.
+For more details, see https://github.com/mobxjs/mobx/issues/532`);
+
 export interface IObservableObject {
 	"observable-object": IObservableObject;
 }