+++ test/array.js
@@ -343,5 +343,57 @@ test('isArrayLike', t => {
 	t.equal(isArrayLike({}), false);
 
 	t.end();
-})
+});
+
+test('.move throws on invalid indices', t => {
+	const arr = [0, 1, 2];
+	const observableArr = observable(arr);
+
+	t.throws(() => observableArr.move(-1, 0), /Index out of bounds: -1 is negative/);
+	t.throws(() => observableArr.move(3, 0), /Index out of bounds: 3 is not smaller than 3/);
+	t.throws(() => observableArr.move(0, -1), /Index out of bounds: -1 is negative/);
+	t.throws(() => observableArr.move(0, 3), /Index out of bounds: 3 is not smaller than 3/);
+
+	t.end();
+});
+
+test('.move(i, i) does nothing', t => {
+	const arr = [0, 1, 2];
+	const observableArr = observable(arr);
+	var changesCount = 0;
+	observableArr.observe(changes => ++changesCount);
+
+	observableArr.move(0, 0);
+
+	t.equal(0, changesCount);
+
+	t.end();
+});
+
+test('.move works correctly', t => {
+	const arr = [0, 1, 2, 3];
+
+	function checkMove(fromIndex, toIndex, expected) {
+		const oa = observable(arr);
+		var changesCount = 0;
+		oa.observe(changes => ++changesCount);
+		oa.move(fromIndex, toIndex);
+		t.deepEqual(oa.slice(), expected, ".move(" + fromIndex + ", " + toIndex + ")");
+		t.equal(changesCount, 1);
+	}
+
+	checkMove(0, 1, [1, 0, 2, 3]);
+	checkMove(0, 2, [1, 2, 0, 3]);
+	checkMove(1, 2, [0, 2, 1, 3]);
+	checkMove(2, 3, [0, 1, 3, 2]);
+	checkMove(0, 3, [1, 2, 3, 0]);
+
+	checkMove(1, 0, [1, 0, 2, 3]);
+	checkMove(2, 0, [2, 0, 1, 3]);
+	checkMove(2, 1, [0, 2, 1, 3]);
+	checkMove(3, 1, [0, 3, 1, 2]);
+	checkMove(3, 0, [3, 0, 1, 2]);
+
+	t.end();
+});